move LockedKernel to locks.rs
