+++ src/gpu/mod.rs
@@ -42,38 +42,3 @@ use ocl::Device;
 lazy_static::lazy_static! {
     pub static ref GPU_NVIDIA_DEVICES: Vec<Device> = get_devices(GPU_NVIDIA_PLATFORM_NAME).unwrap_or_default();
 }
-
-pub struct LockedKernel<K, F>
-where
-    F: Fn() -> Option<K>,
-{
-    _f: F,
-    kernel: Option<K>,
-}
-
-impl<K, F> LockedKernel<K, F>
-where
-    F: Fn() -> Option<K>,
-{
-    pub fn new(f: F) -> LockedKernel<K, F> {
-        LockedKernel::<K, F> {
-            _f: f,
-            kernel: None,
-        }
-    }
-    pub fn get(&mut self) -> &mut Option<K> {
-        #[cfg(feature = "gpu")]
-        {
-            use log::{info, warn};
-            if !PriorityLock::can_lock() {
-                if let Some(_kernel) = self.kernel.take() {
-                    warn!("GPU acquired by a high priority process! Freeing up kernels...");
-                }
-            } else if self.kernel.is_none() {
-                info!("GPU is available!");
-                self.kernel = (self._f)();
-            }
-        }
-        &mut self.kernel
-    }
-}
